Trigger list render only when creating entities
My previous assumption was the entity should be immutable, so
rendering would only happen when replacing an entity with a newer
incarnation on the stream. But since the entity does have a mutable
state at least for its transient period, the entity must be able
to trigger rerendering by itself anyway. Doing this for the source
part as well allows the stream to not change until an entity gets
created or deleted. This avoids rerendering of any componet that
listens only to a single, completely different entity just because
it does that by filtering the source stream.