Don't rely on Proxy being available
The @stencil/store exposes its current state through a Proxy object,
which acts as if it was a plain object in the shape of the source
data, but it actually passes every property access to the internal
event handling. I can easily avoid this because I don't have dynamic
properties anyway but only a single "data" property that holds a
list of every known entity.